Table bigquery architecture (#530)
* integrations done

* running workflows

* intermediate tables

* cleanup

* delete old logic

* add slug for datasets

* empty migration for teams

* migration v1

* fixes

* seems to work now

* almost there?

* rename namespaces

* add unique constraint and fix bq_table naming

* update fixtures

* fix

* more fix

* cleanup team datasets on deletion

* invites test name

* fix teams and invites tests

* navigating to home goes to the selected team

* fix users tests

+++ apps/tables/signals.py
@@ -8,7 +8,10 @@ from .models import Table
 
 
 @receiver(post_delete, sender=Table)
-def delete_img_pre_delete_post(sender, instance, *args, **kwargs):
+def delete_bigquery_table(sender, instance, *args, **kwargs):
+    # don't delete tables for our cypress fixtures
+    if instance.bq_dataset.startswith("cypress_"):
+        return
     try:
         bigquery_client().delete_table(instance.bq_obj)
     except (NotFound):
